+++ docs/changelog.rst
@@ -15,13 +15,12 @@ Added
 - Included new example notebook for basic HydroMT use through CLI. (PR #202)
 - New method **setup_retentions**. (PR #170)
 - support for non-default paths for mdu/network and missing geoms (PR #271)
-- when reading a model prefer crs from network, then geoms, then userdefined (PR #271)
-- prevent reading/writing a model without a mesh (PR #271)
 
 Changed
 -------
 - Dropped support for Python 3.9. (PR #214)
 - Updated to hydromt v1 API and dropped support for Python 3.10. (PR #162)
+- when reading a model prefer crs from network, then geoms, then userdefined (PR #271)
 
 Fixed
 -----
@@ -30,6 +29,7 @@ Fixed
 - fixed crs/bounds/region properties of DFlowFMMeshComponent (PR #251)
 - crs cannot be None in writing mode (PR #264)
 - properly parsing tunit/startdatetime/stopdatetime (PR #267)
+- prevent reading/writing a model without a mesh and prevent reading a model without an mdu file (PR #271)
 
 
 v0.3.0 (18 October 2024)